refactor: P2 server split 1/N — extract /api/jobs route domain + shared http helpers
First route-domain extraction per docs/quality-remediation-plan.md. Pure
cut-paste, zero behavior change:

- server/http-helpers.ts: readBody / json / matchRoute moved verbatim and
  exported; server.ts now imports them
- server/routes/jobs.ts: all 8 /api/jobs* blocks (previously at two sites
  in the giant handler, relative order preserved) become
  handleJobsRoutes(req, res, method, url, ctx): Promise<boolean>; server.ts
  keeps a single delegation line at the original position
- Drop now-unused executeJob/validateCron imports from server.ts
- Smoke test gains two jobs assertions (POST invalid cron -> 400, per-id
  runs) to lock the extracted domain's behavior

server.ts 5295 -> 5155 lines. Pattern for the remaining domains is set.
